Move tests related to deprecation from test_cocotb to test_deprecation
