+++ tests/test_cases/test_cocotb/test_deprecated.py
@@ -1,3 +1,6 @@
+# Copyright cocotb contributors
+# Licensed under the Revised BSD License, see LICENSE for details.
+# SPDX-License-Identifier: BSD-3-Clause
 import cocotb
 import warnings
 from contextlib import contextmanager
@@ -13,7 +16,17 @@ def assert_deprecated():
             yield warns  # note: not a cocotb yield, but a contextlib one!
     finally:
         assert len(warns) == 1
-        assert issubclass(warns[0].category, DeprecationWarning)
+        assert issubclass(warns[0].category, DeprecationWarning), "Expected DeprecationWarning"
+
+
+@contextmanager
+def assert_raises(exc_type):
+    try:
+        yield
+    except exc_type:
+        pass
+    else:
+        raise AssertionError("{} was not raised".format(exc_type.__name__))
 
 
 @cocotb.test()
@@ -36,3 +49,16 @@ async def test_unicode_handle_assignment_deprecated(dut):
     with assert_deprecated() as warns:
         dut.string_input_port <= "Bad idea"
     assert "bytes" in str(warns[0].message)
+
+
+@cocotb.test()
+async def test_create_error_deprecated(dut):
+    with assert_deprecated():
+        _ = cocotb.result.create_error(cocotb.triggers.Timer(1), "A test exception")
+
+
+@cocotb.test()
+async def test_raise_error_deprecated(dut):
+    with assert_deprecated():
+        with assert_raises(cocotb.result.TestError):
+            cocotb.result.raise_error(cocotb.triggers.Timer(1), "A test exception")
